maingameloop.java moved to FeatureTester and software.java renamed as maingame.java for better understanding what each of them mean. since SCALE from both featuretester and maingame is used in other classes it is moved to the separate class since the number of their common variable fields can expand and accidentially cause some troubles.
