+++ GameEngine2/build.gradle
@@ -144,10 +144,10 @@ test {
 }
 
 // Define the main class for the application
-mainClassName = 'MainGame.MainGameLoop'
+mainClassName = 'FeatureTester.FeatureTester'
 
 task(runSoftware, dependsOn: 'classes', type: JavaExec) {
-    main = 'software.Software'
+    main = 'MainGame.MainGame'
     classpath = sourceSets.main.runtimeClasspath
     standardInput = System.in
 }
