+++ GameEngine2/src/main/java/Entities/Camera.java
@@ -1,6 +1,6 @@
 package Entities;
 
-import MainGame.MainGameLoop;
+import MainGame.GameStaticData;
 import MouseHandler.MouseHandler;
 import RenderEngine.DisplayManager;
 import Terrain.Terrain;
@@ -12,7 +12,7 @@ import static org.lwjgl.glfw.GLFW.*;
 
 public class Camera {
 
-    private float distanceFromBall = 5*MainGameLoop.SCALE;
+    private float distanceFromBall = 5*GameStaticData.SCALE;
     private float angleAroundBall = 0;
 
     private Vector3f position;
