+++ GameEngine2/src/main/java/MainGame/GameStaticData.java
@@ -1,0 +1,5 @@
+package MainGame;
+
+public class GameStaticData {
+	static public final int SCALE = 10;
+}
