+++ GameEngine2/src/main/java/OBJConverter/OBJFileLoader.java
@@ -20,7 +20,7 @@ public class OBJFileLoader {
 
     public static ModelData loadOBJ(String objFileName) {
         FileReader isr = null;
-        File objFile = new File("res/models/" + objFileName + ".obj");
+        File objFile = new File("./res/models/" + objFileName + ".obj");
 
         try {
             isr = new FileReader(objFile);
